Move radio functions to separate files
